backend-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
aiosmtplib==3.0.1
python-dotenv==1.0.1
email-validator==2.2.0
python-multipart==0.0.20
aiofiles==25.1.0
